streamlit/requirements.txt
streamlit>=1.33
snowflake-connector-python
pandas
numpy
plotly>=5
altair>=5
seaborn
streamlit-extras
python-dotenv==1.0.1
pyarrow<19.0.0
xlsxwriter
altair
kaleido>=0.2.1
joypy
pywaffle
bar_chart_race
streamlit-lottie
deck.gl